fix!(renovate): move to a standard location (#32)
The renovate GitHub app doesn't like symlinks, which isn't something that running renovate locally catches. This changes the renovate config to simply extend the renovate base config we have.

This limits us because our per-repo renovate config can't be different from the base config for this repo, but trying to link renovate to inherit from a different config file in the same repository has just found a bunch of issues in renovate.